+++ app/src/main/java/net/gsantner/markor/model/AppSettings.java
@@ -231,6 +231,10 @@ public class AppSettings extends GsSharedPreferencesPropertyBackend {
         return getBool(R.string.pref_key__markdown_render_math, false);
     }
 
+    public boolean isMarkdownMermaidEnabled() {
+        return getBool(R.string.pref_key__markdown_render_mermaid, false);
+    }
+
     public List<String> getMarkdownShownYamlFrontMatterKeys() {
         String pref = getString(R.string.pref_key__markdown_always_shown_yaml_front_matter_keys, "title,tags,date");
         List<String> keys = new ArrayList<>(Arrays.asList(pref.replace(" ", "").split(",\\s*")));